feat: :tada: problem solving
This application resolve this task:  A computer starts by printing the numbers 2023, 2024 and 2025. Then, without stopping, it continues printing the sum of the three most recent numbers it has printed, such as 6072, 10121, 18218, and so on. Can you tell me what the last four digits of the number printed in position 2023202320232023 are? To illustrate, in position 50, the printed number is 8188013234823360, which ends in 3360.